requirements.txt
docxtpl==0.6.4
Jinja2==2.11.1
jinja2schema==0.1.4
lxml==4.5.0
MarkupSafe==1.1.1
Pillow==7.1.1
PySimpleGUI==4.18.0
python-docx==0.8.7
